refactor oauth connect queries into separate files
