+++ src/hooks/query/index.ts
@@ -1,0 +1,9 @@
+import { useListInstallationsQuery } from './useListInstallationsQuery';
+import { useListProviderAppsQuery } from './useListProviderAppsQuery';
+import { useOauthConnectQuery } from './useOauthConnectQuery';
+
+export {
+  useListInstallationsQuery,
+  useListProviderAppsQuery,
+  useOauthConnectQuery,
+};
